SpellChecker/requirements.txt
flask
pyspellchecker
sentencepiece
torch
transformers
flask-cors